mediassist-ai/backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
deep-translator
pymongo
passlib[bcrypt]
python-jose
pydantic
email-validator
python-multipart
